Set number of transients per process to integer dtype

+++ software/cadence_whitepaper_plot_pipe.py
@@ -737,15 +737,15 @@ if __name__ == "__main__":
                 id_list3 = None
                 id_list4 = None
 
-            num_trans_pprocess1 = np.ceil(num_detected1/size)
-            num_trans_pprocess2 = np.ceil(num_detected2/size)
-            num_trans_pprocess3 = np.ceil(num_detected3/size)
-            num_trans_pprocess4 = np.ceil(num_detected4/size)
+            num_trans_pprocess1 = int(np.ceil(num_detected1/size))
+            num_trans_pprocess2 = int(np.ceil(num_detected2/size))
+            num_trans_pprocess3 = int(np.ceil(num_detected3/size))
+            num_trans_pprocess4 = int(np.ceil(num_detected4/size))
 
-            receive_array1 = np.empty((int(num_trans_pprocess1), 1))
-            receive_array2 = np.empty((int(num_trans_pprocess2), 1))
-            receive_array3 = np.empty((int(num_trans_pprocess3), 1))
-            receive_array4 = np.empty((int(num_trans_pprocess4), 1))
+            receive_array1 = np.empty((num_trans_pprocess1, 1))
+            receive_array2 = np.empty((num_trans_pprocess2, 1))
+            receive_array3 = np.empty((num_trans_pprocess3, 1))
+            receive_array4 = np.empty((num_trans_pprocess4, 1))
 
             comm.barrier()
             comm.Scatter([id_list1, num_trans_pprocess1, MPI.DOUBLE],
